requirements.txt
alabaster==0.7.12
amqp==5.1.1
asgiref==3.5.2
autopep8==1.6.0
Babel==2.10.3
billiard==3.6.4.0
bleach==5.0.1
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
commonmark==0.9.1
coverage==6.4.3
distlib==0.3.4
dj-database-url==1.0.0
Django==4.0.6
django-archive==0.2.0
django-ckeditor==6.5.0
django-cors-headers==3.13.0
django-downloadview==2.3.0
django-email-service==0.3
django-heroku==0.3.1
django-js-asset==2.0.0
django-nested-admin==3.3.0
django-stubs==1.12.0
django-stubs-ext==0.5.0
django-tinymce==3.5.0
docutils==0.19
factory-boy==3.2.1
Faker==14.1.1
filelock==3.7.1
gunicorn==20.1.0
idna==3.3
image==1.5.33
imagesize==1.4.1
importlib-metadata==4.12.0
Jinja2==3.1.2
keyring==23.8.1
kombu==5.2.4
mailjet-rest==1.3.3
MarkupSafe==2.1.1
mypy==0.971
mypy-extensions==0.4.3
packaging==21.3
Pillow==9.2.0
pipenv==2022.6.7
pkginfo==1.8.3
platformdirs==2.5.2
prompt-toolkit==3.0.30
psycopg2==2.9.3
pycodestyle==2.8.0
pydantic==1.9.1
Pygments==2.12.0
pyparsing==3.0.9
pypng==0.0.20
python-dateutil==2.8.2
python-decouple==3.6
python-monkey-business==1.0.0
pytz==2020.5
pywin32-ctypes==0.2.0
readme-renderer==36.0
recommonmark==0.7.1
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.5.1
six==1.16.0
snowballstemmer==2.2.0
Sphinx==5.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.1
toml==0.10.2
tomli==2.0.1
twine==4.0.1
types-pytz==2022.1.2
types-PyYAML==6.0.11
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.2.0
zipp==3.8.1
